requirements.txt


openai
networkx
matplotlib
python-dotenv
pinecone-client==5.0.1
neo4j==5.0.0
sentence-transformers==2.2.0
torch==2.1.0